Mutation Tests + .NET 8 update (#13)
* chore: Add dotnet tool for mutation testing

* chore: Correct task to execute stryker

* chore: Correct path to sln file in execute stryker

* chore: Correct sln filepath for stryker exec

* Fix regex pattern in DiceRollContextParser and add test cases

* Fix dice pool alteration logic in ShadowRunRollContextParser

* Update dotnet-stryker version to 3.12.0 and ShadowRunRollContextParser logic to match mutation testing

* Refactor hit limit calculation in ShadowRunRollResult class

* Add Stryker mutation testing with HTML and dashboard reporters

* Update target frameworks to .NET 8 and package versions

* Refactor InvalidAmountSidesException class

* Refactor dice roll context and die tests

* Refactor dice initialization in ShadowRunRollContextTests

* Remove unnecessary comments

* Add debug configuration to build command to enable sourcelink

* Added GitHub sourcelink package and other necessary properties

* feat: Managed to get mutator report to dashboard working

* chore: Remove dotnet tools file to clear confusion

* chore: Updated files based on SonarCloud maintainable report

* chore: Update next set of SQ recommendations

+++ src/ShadowRoller.Domain/Contexts/ShadowRun/ShadowRunRollContextParser.cs
@@ -19,11 +19,11 @@ public class ShadowRunContextParser : IRollContextParser<ShadowRunRollContext, S
     private static void AlterDicePool(string arg, List<Die> dice)
     {
         var amountDice = GetValue(arg);
-        if (!amountDice.HasValue)
+        if (!amountDice.HasValue || amountDice.Value == 0)
             return;
-        if (amountDice.Value >= 0)
+        if (amountDice.Value >= 1)
             dice.AddRange(Enumerable.Repeat(new Die(AMOUNTSIDES), amountDice.Value));
-        else
+        if (amountDice.Value <= -1 && dice.Count != 0)
             dice.RemoveRange(0, Math.Abs(amountDice.Value));
     }
